split tokens from implementation
