+++ src/packages/user/current-user/index.ts
@@ -3,4 +3,5 @@ export * from './components/index.js';
 export * from './history/current-user-history.store.js';
 export * from './utils/index.js';
 export * from './current-user.context.js';
+export * from './current-user.context.token.js';
 export type * from './types.js';
