+++ src/packages/user/current-user/modals/current-user/current-user-modal.element.ts
@@ -1,4 +1,4 @@
-import { UMB_CURRENT_USER_CONTEXT } from '../../current-user.context.js';
+import { UMB_CURRENT_USER_CONTEXT } from '../../current-user.context.token.js';
 import type { UmbCurrentUserModel } from '../../types.js';
 import { UMB_AUTH_CONTEXT } from '@umbraco-cms/backoffice/auth';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
