+++ src/packages/user/current-user/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_CURRENT_USER_REPOSITORY_ALIAS = 'Umb.Repository.CurrentUser';
