+++ src/packages/user/current-user/repository/current-user.store.ts
@@ -3,10 +3,10 @@ import type {
 	UmbCurrentUserMfaProviderModel,
 	UmbCurrentUserModel,
 } from '../types.js';
+import { UMB_CURRENT_USER_STORE_CONTEXT } from './current-user.store.token.js';
 import type { UmbUserDetailModel } from '@umbraco-cms/backoffice/user';
 import { UMB_USER_DETAIL_STORE_CONTEXT } from '@umbraco-cms/backoffice/user';
 import { UmbContextBase } from '@umbraco-cms/backoffice/class-api';
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UmbArrayState, UmbObjectState } from '@umbraco-cms/backoffice/observable-api';
 
@@ -106,5 +106,3 @@ export class UmbCurrentUserStore extends UmbContextBase<UmbCurrentUserStore> {
 }
 
 export default UmbCurrentUserStore;
-
-export const UMB_CURRENT_USER_STORE_CONTEXT = new UmbContextToken<UmbCurrentUserStore>('UmbCurrentUserStore');
