+++ src/packages/user/current-user/repository/index.ts
@@ -1,3 +1,4 @@
+export { UMB_CURRENT_USER_REPOSITORY_ALIAS } from './constants.js';
+export { UMB_CURRENT_USER_STORE_CONTEXT } from './current-user.store.token.js';
 export { UmbCurrentUserRepository } from './current-user.repository.js';
-export { UMB_CURRENT_USER_REPOSITORY_ALIAS } from './manifests.js';
-export { UmbCurrentUserStore, UMB_CURRENT_USER_STORE_CONTEXT } from './current-user.store.js';
+export { UmbCurrentUserStore } from './current-user.store.js';
