+++ src/packages/user/current-user/utils/is-current-user.function.ts
@@ -1,4 +1,4 @@
-import { UMB_CURRENT_USER_CONTEXT } from '../current-user.context.js';
+import { UMB_CURRENT_USER_CONTEXT } from '../current-user.context.token.js';
 import { UmbContextConsumerController } from '@umbraco-cms/backoffice/context-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
