+++ src/packages/user/modals/change-password/change-password-modal.element.ts
@@ -1,5 +1,5 @@
-import { UmbUserItemRepository } from '../../user/repository/item/user-item.repository.js';
-import { UMB_CURRENT_USER_CONTEXT } from '../../current-user/current-user.context.js';
+import { UMB_CURRENT_USER_CONTEXT } from '@umbraco-cms/backoffice/current-user';
+import { UmbUserItemRepository } from '@umbraco-cms/backoffice/user';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
 import type { CSSResultGroup } from '@umbraco-cms/backoffice/external/lit';
 import { css, html, nothing, customElement, state } from '@umbraco-cms/backoffice/external/lit';
